backend/foodgram/requirements.txt
asgiref==3.6.0
Django==2.2.19
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.0
flake8==5.0.4
gunicorn==20.0.4
importlib-metadata==4.2.0
mccabe==0.7.0
Pillow==8.3.1
psycopg2-binary==2.8.6
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.6.0
python-dotenv==0.21.0
pytz==2022.6
sqlparse==0.4.3
typing_extensions==4.4.0
zipp==3.11.0